move asset type
