+++ packages/asset-uploader/src/delete.ts
@@ -1,11 +1,12 @@
 import { DeleteObjectCommand } from "@aws-sdk/client-s3";
-import { prisma, Asset } from "@webstudio-is/prisma-client";
+import { prisma } from "@webstudio-is/prisma-client";
 import { unlink } from "fs/promises";
 import path from "path";
 import { deleteAssetInDb } from "./db";
 import { imageFsDirectory } from "./utils/image-fs-path";
 import { s3EnvVariables } from "./schema";
 import { getS3Client } from "./targets/s3/client";
+import { Asset } from "./types";
 
 export const deleteAsset = async ({
   id,
