+++ packages/asset-uploader/src/index.ts
@@ -2,3 +2,4 @@ export * from "./upload";
 export * from "./delete";
 export * from "./db";
 export * from "./utils/format-asset";
+export * from "./types";
