+++ packages/asset-uploader/src/utils/format-asset.ts
@@ -1,4 +1,5 @@
-import { Asset, DbAsset } from "@webstudio-is/prisma-client";
+import { DbAsset } from "@webstudio-is/prisma-client";
+import { Asset } from "../types";
 import { getAssetPath } from "./get-asset-path";
 
 export const formatAsset = (asset: DbAsset): Asset => {
